Swagger: Added initial description of the current api
The json tags were adjusted slightly as some of them was not
interpretable and some used underscore-space style. Some of the models
were adjusted to make support for alternate task definitions cleaner in
the future - logic for deferring parsing the task was also added to
accommodate this change.

NOTE: These swagger specs should be filled out more accurately with
      examples and descriptions.

Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>