store: Rework mock and improve error handling
A proper mock is added to the datastore mock, and some of the arguments
to the datastore interface was extended with a context argument where
necessary.
The channel from the job scheduler is made generic (interface{}) such
that the main worker routine is notified when an error occurs.

changelog: none

Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>