Improve error handling of HTTP tasks
Workflow definition support nows an optional list of response status
codes to be considered valid. If the list is populated with at least one
item, the HTTP status code of the response is checked to determine if
the task was successful or not.

If a workflow task fails, the workflow is stopped and the status is set
to StatusFailure.

Changelog: None

Signed-off-by: Fabio Tranchitella <fabio@tranchitella.eu>