Add "cli" task type to run command line-based tasks
We introduce a new "cli" task type, which allows to run command line commands
and capture standard output and standard error. The exit code determines if the
task is successful or not: exit code 0 is succesfull, otherwise the task failed
and the workflow execution is stopped.

This new task type supports an optional timeout to limit the execution time of
the command. The default execution time is 4 hours.

Changelog: none
Signed-off-by: Fabio Tranchitella <fabio@tranchitella.eu>