Avoid race condition when starting a workflow
Change the order of insertion of the Job in the queue and generic jobs
collections. If we insert the job in the queue before inserting it into
the generic jobs collection, we can have a race condition when a worker
picks up the job before the server completed the insertion.

Changelog: none

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>