[pipeline] Use template to build/publish Docker images
Now that the template supports defining a custom Dockerfile, we can use
it to build/publish workflows-worker image in addition to the server
one.

The main motivation is to keep alignment with the rest, easing
maintenance when introducing changes in our publish policies (i.e.
if/when we want to publish feature/* based images).

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>