New properties: `requires` for Tasks, `skippped` for TaskResult
We introduce a new property for Tasks named `requires`; it contains a
list of strings, eventually containing input parameters. env variables
or golang templates that will be interpolated before evaluation. If
requires is set, all the values of the list must evaluate to a non-empty
string to executed the task, otherwise it will be skipped.

Changelog: title

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>