ci: Fix acceptance tests coverage not being reported
The CI coverage is not being reported for acceptance tests because of
permission errors. This commit refactors the coverage generation to use
`go build -cover` instead of compiling a custom test. The coverage
reported from acceptance tests will be put in the directory
`tests/coverage` directory (which MUST have 0777 permission bits for
"nobody" to write there). The aggregated coverage can be aggregated and
converted to text format using the built-in `go tool covdata`.

Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>